pyproject.toml
[project]
name = "wechat-mp-skill"
version = "0.1.0"
description = "微信公众号技能 - Claude Code skill for WeChat Official Account management"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
    { name = "helebest" }
]
keywords = ["wechat", "weixin", "公众号", "claude-code", "skill"]

dependencies = [
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.1.0",
    "responses>=0.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "e2e: end-to-end tests requiring real API credentials",
    "unit: unit tests with mocked responses",
]

[tool.coverage.run]
source = ["scripts"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
]

[tool.hatch.build.targets.wheel]
packages = ["scripts"]
